Cost_Prediction/requirements.txt
azure-functions
pandas==1.3.4
scipy==1.5.4
scikit-learn==0.23.2
numpy==1.21.3
numba==0.53.1
pyodbc==4.0.30